Reorganize into slackbot_workout folder
